// src/app/user/models/user.ts
  export class Semester {
    semester?: string;
    gpa?: number;
    courses?: Array<any>;
  }
  

  export class Transcript {
    type?: string;
    mobile?: string;
    email?: string;
    cnic?: string;
  }

  export class Status {
    status?: string;
  }
  
  export class User {
    name?: string;
    password?: string;
    email?: string;
    role?: string;
    enrollment?: string;
  }

  
  export class Employee {
    id?: number;
    name?: string;
    password?: string;
    email?: string;
  }
  